--- a/PrimeRust/solution_1/.gitignore
+++ b/PrimeRust/solution_1/.gitignore
@@ -1,1 +1,2 @@
 target/
+perf*.data