--- a/PrimeRust/solution_1/perf-collect
+++ b/PrimeRust/solution_1/perf-collect
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+cargo build --release
+perf record target/release/prime-sieve-rust -t 1 --bits-unrolled -s 15