--- a/PrimeRust/solution_1/perf-report
+++ b/PrimeRust/solution_1/perf-report
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+perf report -M intel