--- a/PrimeRust/solution_1/src/main.rs
+++ b/PrimeRust/solution_1/src/main.rs
@@ -898,16 +898,16 @@ fn main() {
                     opt.print,
                 );
             }
-            for _ in 0..repetitions {
-                run_implementation::<FlagStorageUnrolledBits32>(
-                    "bit-storage-unrolled32",
-                    1,
-                    run_duration,
-                    threads,
-                    limit,
-                    opt.print,
-                );
-            }
+            // for _ in 0..repetitions {
+            //     run_implementation::<FlagStorageUnrolledBits32>(
+            //         "bit-storage-unrolled32",
+            //         1,
+            //         run_duration,
+            //         threads,
+            //         limit,
+            //         opt.print,
+            //     );
+            // }
         }
     }
 }