@GordonBGood, here's the first cut :) It's working, but not working very well just yet.

Some issues to resolve still: in particular, the unrolled indices are not getting pulled in as constants yet, so the compiler isn't able to coalesce these operations on the same word into a single load, or, or, or, store. I'm fighting with the compiler a bit.

So far I haven't had to resort to procedural macros, but that will be the next step if I can't find a workaround to the above problem. I discovered it's going to be a little messy in the context of my existing solution because they'll need to reside in a separate crate, and I'll need to move the whole project structure around to accommodate it. 

The other relevant issue is that I probably should be using specific reset functions for the sparse resets too. And I'll need to fix the above problem to get that working correctly too. Right now I'm just using a generalised sparse reset function.

Considering the above issues still to be resolved, the performance is looking pretty good: around 22k iterations on the 8b implementation. There's light at the end of the tunnel :)

I really like your Chapel implementation. Especially using the 64b words for the dense part, but retaining the 8b version for the sparse resets :) That is a very neat idea. And it's fast as anything on my machine -- nearly 30k iterations vs. about 25k for my existing striped blocks hybrid.