@GordonBGood 2021-08-23
@mike-barber:

> @GordonBGood, here's the first cut :) It's working, but not working very well just yet.
> 
> Some issues to resolve still: in particular, the unrolled indices are not getting pulled in as constants yet, so the compiler isn't able to coalesce these operations on the same word into a single load, or, or, or, store. I'm fighting with the compiler a bit.
> 
> So far I haven't had to resort to procedural macros, but that will be the next step if I can't find a workaround to the above problem. I discovered it's going to be a little messy in the context of my existing solution because they'll need to reside in a separate crate, and I'll need to move the whole project structure around to accommodate it.

Yes, that's one of the reasons I haven't used Rust's procedural macros - the need to put them in a separate crate; it's awkward but not impossible and is the same as GHC Haskell using Template Haskell - the templates/macros need to be in a separate module/separate file.

Rust's `macro-rules` templates can fix the problem of creating literals, but they can't generate variable code as I believe is required for the "dense" solution.  Here is an [Rust example of using `macro_rules` to generate the extreme loop unrolling for a Sieve of Eratosthenes benchmark that just measures the time to cull a CPU L1 data cache buffer multiple times](https://play.rust-lang.org/?version=stable&mode=release&edition=2018&gist=7a63378e511fff6ab5bce0dcc346a24f).  Checking the generated assembly, we get the following for a unrolled loop pattern (for base prime three starting at index of nine, which is bit index of three):
```rust
.LBB25_4:
	orb	$8, (%rdi,%rcx)
	orb	$64, (%rdi,%r9)
	orb	$2, (%rdi,%r10)
	orb	$16, (%rdi,%r11)
	orb	$-128, (%rdi,%r14)
	orb	$4, (%rdi,%r15)
	orb	$32, (%rdi,%r12)
	orb	$1, (%rdi,%r13)
	addq	%rsi, %rdi
	leaq	(%rdi,%rcx), %rax
	cmpq	%rbx, %rax
	jb	.LBB25_4
```
Note the unneeded `leaq` conflagration as part of the loop limit check, which is not in other better optimized versions from other languages; I think that this is a missed optimization on LLVM's part.  There is also that the Rust compiler authors admit themselves that they don't do a good job of optimizing their LLVM IR generation and depend on LLVM to optimize away the bloat, so it's also possible that this is an artifact of what LLVM was given and this was it's best solution.  **Hold the fort!**  I just looked at the generated LLVM IR code and this is an artifact of what Rust hands LLVM - it has inlined the generation of the limit pointer used as a limit check, so is generating it anew on every loop!  This should have been recognized and lifted out of the loop, if not by the Rust compiler then by LLVM.

Anyway, it won't make much difference, as it will only add about a third of a CPU clock cycle for every about eight CPU clock cycles or about four percent, but it means that this code will lose out to Nim, which doesn't have it.  As the Nim back end is C or C++ code to be finally compiled by either gcc or clang, this also means if someone takes the trouble to implement my technique in C or C++ they also won't have this overhead (as long as Loop Strength Reduction is turned off for clang, which error adds even worse overhead of a different kind).

> The other relevant issue is that I probably should be using specific reset functions for the sparse resets too. And I'll need to fix the above problem to get that working correctly too. Right now I'm just using a generalised sparse reset function.

Yes, that's why you will need procedural macros unless you are willing to do use external code generation as I did in the Chapel version as Chapel doesn't have macros.

> Considering the above issues still to be resolved, the performance is looking pretty good: around 22k iterations on the 8b implementation. There's light at the end of the tunnel :)
> 
> I really like your Chapel implementation. Especially using the 64b words for the dense part, but retaining the 8b version for the sparse resets :) That is a very neat idea. And it's fast as anything on my machine -- nearly 30k iterations vs. about 25k for my existing striped blocks hybrid.

Sounds like your machine is about as fast as mine - that's what I get for my Chapel version - almost 30K passes in five seconds.  But like I said in the other thread, the Nim version is likely to be faster yet again as I can use macros (Nim's macros can be in the same file as where they are used).

I think when fully implemented, your speed should be at least as fast as my latest Chapel version, as Chapel doesn't optimize to the maximum limit either.