@GordonBGood 2021-08-23
@mike-barber:

Now that my Chapel "unrolled_hybrid" version has wiped the board, there still is a consideration of the differences of how CPU's handle "cache thrashing":  For modern CPU's and even the lowly Raspberry Pi 4's arm64 Cortex A72 my densely packed "Extreme Loop Unrolling" functions work at least a little better than "unpeeled"/"striped" by blocks or as you call "general blocked", but for the J1800 Celeron and its ilk, the latter is faster.  I think this is due to the way the cache lines are handled by this truly primitive CPU.  So if I had a different version using "unpeeled block" instead of "unrolled" for the sparse 8-bit culling, my J1800 score of almost 10K might be 11K or 12K, just for the Celeron J1800, but would be worse than the current version on the other two.

I don't know that I'll do this for Chapel as I only introduced it to give it whatever publicity it can gain from the race, but I'm pretty sure I will introduce the variations between the hybrid combinations for Nim and you might want to do the same for Rust.