@GordonBGood 2022-06-19
Hi @mike-barber:

I've been thinking that we should write some sample code in the new language, which looks like may be named Fir (repo as fir-lang) in order to see what code looks and "feels" like, and although we don't have a working compiler yet, we know what the C code that we hope to generate should look like so will know its eventual approximate speed, with the goal to make it as as the same speed as Nim (and faster or slower if compiled with clang rather than GCC depending on the actual code).

As to "type classes"/"abilities", GHC Haskell has a very nice extension called "FlexibileContexts" that means that one doesn't have to list all the required "abilities" in a function type signature but they are derived automatically when one uses the associated "interface" functions on the given type (a bit of a return to "duck typing" but applied to use rather than definition).  From what I read, this extension doesn't really have any downsides other than making code a little more short hand to type and not being quite so explicit as to "abilities" used, where the compiler would complain if an "ability" was ambiguous and couldn't be derived (which shouldn't happen as long as "ability" functions have unique names).  I think we should use the effect of having this extension turned on all the time, which doesn't preclude the programmer using the context requirements in the type signatures anyway if they want to be more explicit.  Another advantage of using it in being backward compatible with Elm code is that the predefined Elm contexts of `number`, `comparable`, `appendable`, and `comappend` can be ignored and become just normal type variable placeholders with the right "ability/abilties" applied automatically depending how it is used in the value/function definition by "FlexibleContexts".  As to giving ability functions unique names, we can control that if we define all the abilities we think are required and don't allow non-core-Fir developers to add to them just as currently in Elm for symbolic operators, effect managers, and native "Kernel" code libraries (other than as channeled through ports), and I think we may want to keep those restrictions, but we'll see...

There is one other consideration we need to make:  In Haskell with its richness of primitive integer and float types, there are defaults for the automatic types assigned if there are no other hints that can be inferred; thus, if a literal number has a decimal point then it must be a `Float` and a Double is assumed by default (with a warning if turned on) and a integer literal is automatically assumed to be a multi-precision `Integer` - however, if one wants to override the default, one can apply a type signature just to the value.  This last is not possible in Elm, where type signatures are for value/function declarations at the global or local level (with a `let ... in`) so we can't force a type without creating a named value and using that, which is another LOC; and when we add all of the desired primitive types of different bitedness and signedness, we may quite frequently need this - for instance if we are creating an intermediate array of `Byte`, there may be no other way to force the type inference to know that it has `Byte` elements other than by giving a value somewhere, perhaps in initialization.  We can eliminate the extra required LOC in two ways:  1) allowing an internal type signature to individual local literals as in Haskell, or 2) by using a primitive number literal suffix to indicate the type, where the type is to be inferred if the suffix is not used.  We will need to think about this and decide if this is necessary and, if so, which method to use...

So what sample code?  Well, since we are working here from "The Software Drag Race", we may as well start with that, although later we will need some more benchmarks to show the efficiency of automatic reference counting as compared to garbage collection languages and also no memory management (ugly as that generally is).  However, there is no pressing need for this latter as we know the general speed of reference counting as used in Rust, etc., and when completed, ours should be faster than that as it is done "behind the scenes" and the compiler will be able to elide away many cases of needing to count references.  Also, the BDFL of Nim has done such a study showing the automatic reference counting is generally faster, and using similar techniques, our should be at least as fast as his...

I've been thinking of the nature of computer language syntax, especially in the case of Rust which seems to be a language that is going to "make it" in that if fills a need for a low level language that is safer than C/C++, yet as powerful and about as fast.  I briefly considered using it as our compilation target but turned away from it because our "safety" will be provided by our own pure functional syntax without the programmer needing to be concerned with all the details of borrowing (which only very occasionally happens in Fir and only in an monadic context) and lifetimes (which a Fir programmer doesn't need to concern themselves with as everything is handled by Data Flow Analysis graphing by the Fir compiler), it gives less compiler target options (GCC vs. clang vs. whatever aren't options), and how slow it is to compile.  I think that Fir may very well give all the advantages of Rust without all the complexity represented to the programmer because of its use of the FP paradigm, although the programmer will have to adapt to that...

So, a "Software Drag Race" solution written in Elm imagining that it has been extended to emit C as per Fir is in the following code; since "abilities" are more of a convenience so that one doesn't have to continuously concern themselves with module imports and whether there are naming conflicts between modules, we can write it with just unique names and modules in existing Elm, with comments where the actual implementation would differ:
```elm
module Main exposing (main)

-- The Console module is automatically imported by Prelude;
-- a terminal always opens unless a compiler switch as `--no-console` is used:
-- import Console exposing (outStringLine, outString, inStringLine, inChar,
--                          errorStringLine, errorString)
-- with side effects, they must generate `Task`'s with signatures as follows:
-- out.../error... : String -> Task Never ()
-- inStringLine : () -> Task Never String -- these work like `now` from `Time`
-- inChar : () -> Task Never Char

-- the SafeIndex module provides a set of functions and macros that can be used to
-- provide pre-checked "lifted" guarantees that an index is within bounds, and also generate the
-- "boilerplate" code to group operations in various powers of two sized groups with provision for
-- "custom" variants where one cant specify the derivation of compiler generated constant values to
-- write or modify an indexed types contents...
import SafeIndex exposing (..) -- names in this module are unique enough for unlikely clashes!
import Dict exposing (Dict)
-- should be LinearArray instead of Array with more Int bitedness types...
import Array exposing (Array)
import Task exposing (Task)
import Time exposing (now, posixToMillis)

type alias Prime = UInt64

cLIMIT : Prime
cLIMIT = 1_000_000

cTIME : Int
cTIME = 5_000

cEXPECTED : Dict Int Int
cEXPECTED = [ (10, 4) , (100, 25), (1_000, 168), (10_000, 1229)
            , (100_000, 9592), (1_000_000, 78_498), (10_000_000, 664_579)
            , (100_000_000, 5_761_455), (1_000_000_000, 50_547_543)
            ] |> Dict.fromList

-- just a place holder for the real thing...
-- we maintain `Array` for backward compatibility
type alias LinearArray a = Array a

-- kind of emulates what the `ToStringable` ability will do...
-- there will also be a `FromStringable` ability: converts a String to a type...
toString : a -> String
toString = Debug.toString

-- kind of emulates what the real Console module function will do...
outStringLine : String -> Task Never ()
outStringLine str = let _ = Debug.log "" str in Task.succeed ()

-- following is compiled as a C `enum`; is part of the `Enumerable` ability...
type
  -- the `Enumerable` ability is the same as the Haskell `Enum` type class and
  -- also the `Boundable` ability as per the Haskell `Bounded` type class...
  Technique = BitTwiddle | Stride8 | Stride8Block | Extreme | ExtremeHybrid
  -- has Enumerable, Boundable;
  -- may have Equable, Comparable, ToStringable, FromStringable...

cMASKS : LinearArray UInt8
cMASKS = [| 1, 2, 4, 8, 16, 32, 64, 128 |] -- new literal linear array

-- macro that does extreme unrolling of stride8 loop...
-- marr_u8 is a `ChangingLinearArray s Uint8`, strt/stop/step are
-- `Int` start/limit bit positions and `Int` step value, respectively...
macro unrollSetBits marr_u8 strt stop step =
  -- following takes a step `Int` constant
  -- a unrolled loop index `Int` constant and
  -- the element before modification, and then
  -- returns a the new value of the element using a
  -- compiler generated constant operation...
  {% range stp ni si %}
    safeIndexedRead {{si}}
    |> andThen ( \ oldv ->
        safeIndexedWrite {{si}}
          (oldv ||| (1 <<< (((stp * stp - 3) >>> 1) + ni * stp) &&& 7)) )
  let swtch : UInt8
      swtch = fromIntegral <| (bp &&& 6) >>> 1 in
  |> andThen ( \ _ ->
      case swtch of
        {% range n 0 63 %}
          (n) ->
            -- the following is a macro itself,
            -- so it can be passed a MACRO_CONSTANT...
            safeIndexedModifySequenceByDiv8CustomUnrollBy
              8 marr_u64 {{step}} {{strt}} {{stop}} TABLE_MASKS unroll
            |> Sequence.reduceFrom (wrap ()) (always << andThen wrap)
      _ -> error "Should never get here!"
      |> andThen (
          let ns0 = (unrollLmt + 1 - {{strt}}) // {{step}} * {{step}} + {{step}}
              cmsks = safeIndexedGetSequenceByMod cMASKS {{step}} ns0 {{stop}}
          in
          safeIndexedModifySeqByDiv8
            {{marr_u8}} {{step}} ns0 {{stop}} cmsks (|||)
          |> Seq.reduceFrom  (wrap ()) (always << andThen wrap) ) )

-- macro that does extreme unrolling of dense culling loop by
-- culling within a 64 bit word until complete, storing,
-- loading the next word, and repeating for
-- 64 compiler computed bit patterns...
-- marr_u8 is a `ChangingLinearArray s Uint8`, strt/stop/step are
-- `Int` start/limit bit positions and `Int` step value, respectively...
macro denseCullsSetBit marr_u8 strt stop step =
  unsafeCastLinearArray marr_u8
  |> andThen ( \ marr_u64 ->
      let dnsstrt = let loop i = if i &&& 63 == 0 then i else
                             loop (i + {{step}}) in loop {{strt}}
          imsks = safeIndexedGetSequenceByMod cMASKS {{step}} {{strt}} (dnsstrt - 1)
          swtch : UInt8
          swtch = fromIntegral <| (({{step}} - 3) &&& 127) >>> 1 in
      sameIndexedModifySequenceByDiv8 {{marr_u8}} {{step}} {{strt}} (dnsstrt - 1) imsks (|||)
      |> Sequence.reduceFrom (wrap ()) (always << andThen wrap)
      case swtch of
      -- 64 cases...
      {% range n 0 63 %}
        {{n}} ->
          -- following is a macro itself, so can be passed a MACRO_CONSTANT...
          safeIndexedModifySequenceByDiv64CustomUnrollBy
            8 marr_u64 {{step}} dnsstrt {{stop}} TABLE_MASKS denseUnroll
          |> Sequence.reduceFrom (wrap ()) (always << andThen wrap)
      _ -> error "Should never get here!"
      |> andThen (
          let ns0 = (unrollLmt + 1 - {{strt}}) // {{step}} * {{step}} + {{step}}
              cmsks = safeIndexedGetSequenceByMod cMASKS {{step}} ns0 {{stop}}
          in
          safeIndexedModifySequenceBy {{step}} ns0 {{stop}} cmsks (|||)
          |> Sequence.reduceFrom  (wrap ()) (always << andThen wrap) ) )

-- this is long mostly because it contains two implementations;
-- one using common recursive loops, where with unsafe reads and writes, and
-- one using my "SafeIndex" type ideas where it is obvious that the
-- bounds have been pre-bounds-checked and thus doesn't need UNSAFE blocks...
sievePrimes : Prime -> Technique -> LinearArray UInt8
sievePrimes limit tec =
  if limit < 3 then LinearArray.empty else
  -- the following shows how this can be done when
  -- we have the full set of abilities...
  let lmt = fromIntegral <| (limit - 3) // 2
      sz = max 0 <| ((lmt + 64) // 8) &&& -8 in -- round up to 64-bit boundary
  runChange <| do -- do syntax over the Changing monadic chain...
    -- these `newArray` functions produce array's inited to default zero...
    marr <- newArray_ sz -- marr is (IndexChange s (LinearArray UInt8))
    -- following for inter page cull indexes
    msarr <- newArray_ 8 -- msarr is (IndexChange s (LinearArray Int))
    -- the following is the recursive function loops version...
    let loopbp bp = -- type signature:  Int -> Changing s (LinearArray UInt8)
          if bp * bp > limit then unsafeFreeze marr else
          let i = (bp - 3) / 2
              v = {-# UNSAFE #-} unsafeIndexedRead marr (i >>> 3)
              msk = {-# UNSAFE #-} unsafeIndexedGet cMASKS (i &&& 7) in
           if v &&& msk /= 0 then loopbp (bp + 2) else
          let sqri = (i + i) * (i + 3) + 3 in
          -- nested recursive loops need to be better than current Elm's, which
          -- can only tail call recurse within the same function, not inter function...
          case tec of
          BitTwiddle ->
            let loopc c =
                  -- NOTE that this inner "loop" recurses to an outer "loop" and
                  -- THUS the whole recursion must be "lifted lambda" into loop code...
                  if c > lmt then loopbp (bp + 2) else do
                    let w = c >>> 3 in do
                      {-# UNSAFE #-}
                        v <- unsafeIndexedRead marr w
                        unsafeIndexedWrite marr w (v ||| (indexedGet cMASKS (c &&& 7)))
                      loopc (c + bp)
            in loopc sqri
          Stride8 ->
            let slmt = sqri + 7 * bp
                loops s =
                  if s > slmt then loopbp (bp + 2) else
                  let msk = unsafeIndexedGet cMASKS (s &&& 7) in
                      loopw w =
                        if w > sz then loops (s + bp) else do
                          {-# UNSAFE #-}
                            v <- unsafeIndexedRead marr w
                            unsafeIndexedWrite marr w (v ||| msk)
                          loopw (w + bp)
                  in loopw (s >>> 3)
            in loops sqri
          Stride8Block -> do
            Seq.range 0 7 -- use of zero cost iterator here; not recursion...
            |> Seq.reduce ( \ i ->
                 let s = s0 + i * bp
                     ndx = s &&& 7 in
                 {-# UNSAFE #-} unsafeIndexedWrite msarr ndx s ) (wrap ())
            let bp2 = bp + bp
                bp3 = bp2 + bp
                bp4 = bp3 + bp
                looppg pgi =
                  if pgi >= sz then loopbp (bp + 2) else
                  let loopi i =
                        if i > 7 then looppg (bpi + 16384) else do
                        s <- {-# UNSAFE #-} unsafeIndexedRead msarr i
                        let pglmt = min sz (pgi + 16384)
                            pgbtlmt = (pglmt <<< 3) - 1
                            lmt4w = pglmt - bp3
                            msk = {-# UNSAFE #-} unsafeIndexedGet cMASKS (s &&& 7)
                            loop4w w =
                              if w >= lmt4w then
                                let loopc c =
                                      if c > pgbtlmt then do
                                        {-# UNSAFE #-} unsafeIndexedWrite msarr i c
                                        loopi (i + 1) else do
                                          let w = c >>> 3 in
                                          {-# UNSAFE #-}
                                            v <- unsafeIndexedRead marr w
                                            unsafeIndexedWrite
                                              marr w (v ||| (unsafeIndexedGet cMASKS (c &&& 7)))                                          
                                          loopc (c + bp)
                                in loopc ((w <<< 3) + i) else do
                                  {-# UNSAFE #-}
                                    v0 <- unsafeIndexedRead marr w
                                    unsafeIndexedWrite marr w (v0 ||| msk)
                                    v1 <- unsafeIndexedRead marr (w + bp)
                                    unsafeIndexedWrite marr (w + bp) (v1 ||| msk)
                                    v2 <- unsafeIndexedRead marr (w + bp2)
                                    unsafeIndexedWrite marr (w + bp2) (v2 ||| msk)
                                    v3 <- unsafeIndexedRead marr (w + bp3)
                                    unsafeIndexedWrite marr (w + bp3) (v3 ||| msk)
                                  loop42 (w + bp4)
                        in loop4w (s >>> 3)
                  in loopi 0
            in looppg 0
          Extreme -> unrollSetBits marr s0 lmt bp
          ExtremeHybrid ->
            if bp < 131 then denseCullSetBits marr s0 lmt bp
            else unrollSetBits marr s0 lmt bp
    in loopbp 3
    -- alternative "Seq" implementation, includes safe indexing lifting bounds check...
    let seqi = Seq.range 0 ((sqrti limit - 3) // 2)
        msks = safeIndexedGetSequenceMod cMASKS seqi
        cvs = safeIndexedReadSequenceDiv8 marr seqi in
    Seq.for3ReduceFrom (wrap ()) seqi msks cvs <| \ i msk cv -> do
      v <- cv
      if v &&& msk == 0 then wrap () else
      let bp = i + i + 3
          s0 = (i + i) * (i + 3) + 3 in          
      case tec of
        BitTwiddle ->
          let seqc = Seq.rangeBy bp s0 lmt
              mmsks = safeIndexedGetSequenceByMod cMASKS bp s0 mod seqc in
          safeIndexedModifySeqByDiv8 marr bp s0 lmt mmsks (|||)
          |> Seq.reduceFrom (wrap ()) (always << andThen wrap)
        Stride8 ->
          let slmt = min lmt (s0 + 7 * bp)
              seqs = Sequence.rangeBy bp s0 slmt
              mmsks = safeIndexedGetSeqByMod cMASKS bp s0 slmt in
          Seq.for2ReduceFrom (wrap ()) seqs mmsks <| \ s msk ->
            safeIndexedModifySeqByConstant
              marr bp (s >>> 3) (sz - 1) msk (|||)
            |> Seq.reduceFrom (wrap ()) (\always << andThen wrap)
        Stride8Block ->
          let bp2 = bp + bp
              bp3 = bp2 + bp
              bp4 = bp3 + bp
              ndxs = Sequence.range 0 7
              inits = ndxs |> Range.map (\ i -> (sqri + i * bp)) in
          safeIndexedWriteSeq msarr inits -- init intermediate cull index array...
          |> Seq.reduceFrom (wrap ()) (always << andThen wrap)
          |> andthen ( _ -> 
               Seq.rangeBy 16384 0 (sz - 1)
               |> Seq.reduceFrom (wrap ()) ( \ pgi _ ->
                    let pgnxti = min sz (pgi + 16384)
                        pgbtlmt = (pgnxti <<< 3) - 1
                        lmt4w = pgnxti - bp3 - 1
                        nmsks = safeIndexedGetSeq cMASKS 0 7 in
                    safeIndexedModifySeqCustom msarr 0 7 nmsks <| \ msi msk -> do
                      olds <- safeIndexedRead msi
                      let nw0 = SafeIndex.index msi in
                      -- following is a macro to generate the inline
                      safeIndexedModifySeqByConstantUnrollBy 4 marr bp nw0 lmt4w msk (|||)
                      |> Seq.reduceFrom (wrap ()) (always << andThen wrap)
                      |> andThen ( \ _ ->
                           let ns0 = (((lmt4w + 1 - nw0) // bp4 * bp4) <<< 3) + (olds &&& 3)
…